fix(desktop): settings modal + HDR preview for HEIC original in compare
Settings opens as a topbar modal instead of navigating to /desktop/settings, so changing settings mid-task (e.g. switching to HDR JPEG while picking photos) no longer unmounts the new-album/add page and wipes the in-progress draft. The form is extracted into SettingsPanel; the /desktop/settings route stays for direct access.

The compare modal's 'Original' pane lazily renders a small HDR (Ultra HDR JPEG) preview for HEIC sources, so it glows like the HDR compressed result instead of looking flat (its beforeUrl is only an SDR sips preview). Cached per photo id; only the viewed photo pays the sub-second cost, and it falls back to the SDR preview while rendering or on failure.
