+++ src/app/desktop/galleries/[id]/[album]/add/page.tsx
@@ -234,6 +234,7 @@ export default function AddPhotosPage() {
         .map((p) => ({
           id: p.id,
           name: p.original.name,
+          originalFile: p.original,
           // HEIC has no browser-displayable original — its `originalUrl`
           // is raw HEIC bytes the renderer can't paint, so prefer the
           // sips-decoded `preview`. Other formats keep the full-res
